+++ packages/tools/src/types/PolySeg.ts
@@ -3,5 +3,5 @@ import type { Types } from '@cornerstonejs/core';
 export type PolySegConversionOptions = {
   segmentIndices?: number[];
   segmentationRepresentationUID?: string;
-  viewport?: Types.IViewport;
+  viewport?: Types.IStackViewport | Types.IVolumeViewport;
 };