+++ packages/tools/src/utilities/math/line/index.ts
@@ -2,10 +2,12 @@ import distanceToPoint from './distanceToPoint';
 import distanceToPointSquared from './distanceToPointSquared';
 import distanceToPointSquaredInfo from './distanceToPointSquaredInfo';
 import intersectLine from './intersectLine';
+import isPointOnLineSegment from './isPointOnLineSegment';
 
 export {
   distanceToPoint,
   distanceToPointSquared,
   distanceToPointSquaredInfo,
   intersectLine,
+  isPointOnLineSegment,
 };